+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 
 
+## Version 1.0.3 (2019-11-20)
+
+Remove method name in automatic tag if equals to "invoke".
+Add Inner class name in automatic tag.
+
 ## Version 1.0.2 (2019-11-19)
 
 Fix library publish process.